Moved helpers and errors into internal folders in @itwin/cloud-agnostic-core and @itwin/object-storage-core packages
