+++ cloud-agnostic/core/src/internal/Helpers.ts
