+++ storage/core/src/frontend/index.ts
@@ -4,6 +4,5 @@
  *--------------------------------------------------------------------------------------------*/
 export * from "../common";
 export * from "./FrontendInterfaces";
-export * from "./FrontendHelpers";
 export * from "./FrontendStorage";
 export * from "./FrontendStorageDependency";
