+++ storage/core/src/frontend/internal/index.ts
@@ -2,7 +2,4 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-export * from "./Bindable";
-export * from "./Dependency";
-export * from "./DependencyConfig";
-export * from "./Types";
+export * from "./FrontendHelpers";
