+++ storage/core/src/server/index.ts
@@ -3,6 +3,5 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 export * from "./Interfaces";
-export * from "./Helpers";
 export * from "./ServerStorage";
 export * from "./ServerStorageDependency";
